+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   `network`: Add new command to print list of IPv4 addresses
 -   `rack`: Add new command to visualize hardware racks
 
+### Fixed
+
+-   `save`: Do not ignore capacity (float) incl. unit (b, kb, mb, gb, tb)
+
 ##  [0.8][] – 2019-07-31
 
 ☀️☀️☀️ Happy summer time ☀️☀️☀️